pyproject.toml
[tool.poetry]
name = "FastAPI Example"
version = "0.1.0"
description = ""
authors = ["Lior Nev <nevo.lior@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
loguru = "^0.7.2"
uvicorn = "^0.29.0"
ujson = "^5.9.0"
jinja2 = "^3.1.4"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
pre-commit = "^3.7.0"
ruff = "^0.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
